+++ engines/pug.js
@@ -1,13 +1,12 @@
-import * as pug from "../deps/pug.js";
 import TemplateEngine from "./template_engine.js";
 
 export default class Pug extends TemplateEngine {
   cache = new Map();
-  filters = {};
 
-  constructor(site, options = {}) {
-    super(site, options);
-    this.includes = site.src("_includes");
+  constructor(site, engine, options) {
+    super(site);
+    this.engine = engine;
+    this.options = options;
 
     // Update cache
     site.addEventListener("beforeUpdate", () => this.cache.clear());
@@ -17,10 +16,9 @@ export default class Pug extends TemplateEngine {
     if (!this.cache.has(filename)) {
       this.cache.set(
         filename,
-        pug.compile(content, {
+        this.engine.compile(content, {
+          ...this.options,
           filename,
-          basedir: this.includes,
-          filters: this.filters,
         }),
       );
     }
@@ -31,7 +29,8 @@ export default class Pug extends TemplateEngine {
   addHelper(name, fn, options) {
     switch (options.type) {
       case "filter":
-        this.filters[name] = (text, opt) => {
+        this.options.filters ||= {};
+        this.options.filters[name] = (text, opt) => {
           delete opt.filename;
           const args = Object.values(opt);
           return fn(text, ...args);