+++ plugins/pug.js
@@ -1,3 +1,4 @@
+import * as pug from "../deps/pug.js";
 import Pug from "../engines/pug.js";
 import loader from "../loaders/text.js";
 import { merge } from "../utils.js";
@@ -5,14 +6,23 @@ import { merge } from "../utils.js";
 // Default options
 const defaults = {
   extensions: [".pug"],
+  includes: null,
+  options: {},
 };
 
 export default function (userOptions) {
-  const options = merge(defaults, userOptions);
-
   return (site) => {
-    const pug = new Pug(site);
+    const options = merge(
+      { ...defaults, includes: site.includes() },
+      userOptions,
+    );
 
-    site.loadPages(options.extensions, loader, pug);
+    options.options.basedir = options.includes;
+
+    site.loadPages(
+      options.extensions,
+      loader,
+      new Pug(site, pug, options.options),
+    );
   };
 }